requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.12.0
async-generator==1.10
attrs==22.2.0
browserstack-local==1.2.4
browserstack-sdk==1.4.0
certifi==2022.12.7
charset-normalizer==3.0.1
DateTime==5.0
exceptiongroup==1.1.0
execnet==1.9.0
h11==0.14.0
idna==3.4
iniconfig==2.0.0
mysql-connector-python==8.0.32
outcome==1.2.0
packaging==23.0
pluggy==1.0.0
protobuf==3.20.3
psutil==5.9.4
py==1.11.0
PySocks==1.7.1
pytest==6.2.5
pytest-base-url==2.0.0
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-selenium==4.0.0
pytest-variables==2.0.0
pytest-xdist==3.1.0
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-toolbelt==0.10.1
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tenacity==6.3.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
webdriver-manager==3.8.5
wsproto==1.2.0
zope.interface==5.5.2
